Move viewport converting points to terminal
